Update Api has been renamed to ApiClient. Moved Prediction to its own submodule
